Set release version 1.13.M2 and AspectJ 1.9.7.M2
I used "M2" here in order to mimic the AspectJ milestone we depend on.
It does not feel 100% correct, but it does not matter much and helps me
remember the connection.

I chose 1.13 instead of 1.12.7, because a new minor version is justified
with the fork having a new maintainer and new group ID.