Upgrade PlantUML plugin for JDK 16 compatibility
With the retired plugin, UML generation worked on JDK 8 and 11, not on
16 because of internal API access. The new plugin was accidentally
compiled for Java 11+, so there is an issue with JDK 8 now. Until the
upstream ticket is resolved, we need two plugins and two sets of
versions for the plugin and PlantUML itself. This will go away when the
new plugin will have been released with target Java 8.