Fix Maven site scrolling issues with super-wide table containing <pre>
There is a sample builddef.lst file embedded in a 'pre' section for
parameter 'argumentFileName' in goals 'compile' and 'test-compile',
actually defined in base class AbstractAjcCompiler. I shortened the
code, which results in a smaller width. This helps to avoid problems
when scrolling in a browser like Chrome using cursor keys like 'PgDn'.
Before, due to the super-wide column in the key-value table, the page
was also scrolled horizontally in an ugly way.