Replace PlantUMl plugin by unified version running on Java 8-16
I temporarily replace the upstream 'com.github.funthomas424242' version
by a 'dev.aspectj' deployed on Maven Central, which contains bugfixes
from https://github.com/Huluvu424242/plantuml-maven-plugin/pull/31.

This means, that
  - we now have a single plugin running on Java 8-16 instead of two
    between which to switch in profiles,
  - the fixed plugin no longer requires Maven 3.6.3, but runs on all
    Maven versions down to 3.0.5.

As soon as there is an upstream release with my bugfixes, we can switch
back to it. For for now we are fine and have maximum flexibility for
AspectJ Maven.