Improve 'enablePreview' option description
Code review feedback, thanks to @rfscholte. This parameter has its
gotchas, so I think for AspectJ Maven users it is better to pull it into
the API and document it better than actual CLI compilers do than to let
users play with it via the generic 'additionalCompilerArgs' option and
then answer lots of related StackOverflow and GitHub questions without
even being able to point to the documentation ("we told you so").