Breaking change: use Doxia 2 default directory 'target/reports'
After my PR https://github.com/apache/maven-reporting-impl/pull/26 for
https://issues.apache.org/jira/browse/MSHARED-1327 was merged, the
base directory for reports in AbstractMavenReport when calling mojos
stand-alone, i.e. outside a site generation context, no longer defaults
to 'target/site' but now to 'target/reports'. This is good, because
reports for a Maven site can look different from reports for a single
module. As AjcReportMojo extends AbstractMavenReport, we can respect the
new default easily and even save lines of code and documentation,
because we can just inherit the 'outputDirectory' property instead of
overriding and documenting it.