Convert multimodule_strategy.adoc to ADOC
This is the most complex document, because it contains several sections,
embeds a PlantUML diagram and needs syntax highlighting for Java and
XML.

TODO: Find out why syntax highlighting does not work. It looks as if the
CSS is incorrect.

With this change, we can get gid of multimodule_ajc.puml as a separate
resource and plantuml-maven-plugin, because Asciidoctor understands
embedded PlantUML code blocks, if corresponding optional diagram module
is active.