Say goodbye to org.codehaus.mojo:mojo-parent parent POM
Rationale: The parent site.xml could not be completely overridden in a
nice way, e.g. Mojohaus logo. What little parent information was used in
this plugin that was not overridden and part of our own POM already, has
been added to the plugin POM. The same is true for information from the
parent site.xml.